fix(transport): use correct type for max_instances on Windows
The ServerOptions::max_instances() method in newer tokio versions expects
usize instead of u32. Convert the config value appropriately.

+++ crates/mcpkit-transport/src/windows.rs
@@ -503,7 +503,7 @@ impl TransportListener for NamedPipeServer {
         // Create a new pipe instance for this connection
         let pipe = ServerOptions::new()
             .first_pipe_instance(false)
-            .max_instances(self.config.max_instances)
+            .max_instances(self.config.max_instances as usize)
             .in_buffer_size(self.config.in_buffer_size as u32)
             .out_buffer_size(self.config.out_buffer_size as u32)
             .create(&self.config.name)
